PageForm: extract BasePageForm
